requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
duckduckgo-search==6.1.7
google-generativeai==0.5.4
python-dotenv==1.0.1
slowapi==0.1.9
